remove axi error resp
When c910 accesses to Cheshire's config registers, although the instruction is an LW that only loads 4-byte data, since the AXI interconnect is now 16-byte wide (it's the width of the c910 AXI port), the load continues to access the other 12 bytes which go out of address boundaries. It then causes an LSU exception.

Temporarily fix this problem by suppressing the error response from the Cheshire AXI interconnect, as there was no real error. However, this is only a temporary solution. I believe the real solution is when we divide wider AXI requests (e.g. 64-bit, 128-bit) into 32-bit reg bus requests, we need to refer to the AXI length information. We shouldn't always assume that all 32-bit segmented requests are valid.